pyproject.toml
[project]
name = "ha_kumo_ws"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "httpx[http2]>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-socketio[client]>=5.11.0",
    "pydantic>=2.6.0",
    "aiohttp>=3.9.0",
]
